move Richprint to dedicated DisplayManager module
